feat: implementation of error checking
